Remove kas::layout::Visitor
